pyproject.toml
[project]
name = "scryfall-thermal"
version = "0.1.0"
description = "Print random Scryfall creatures to ESC/POS thermal printers."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31.0",
  "Pillow>=10.2.0",
  "python-escpos>=3.0",
  "cairosvg>=2.7.1",
]

[project.scripts]
scryfall-thermal = "scryfall_thermal.main:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
